Crate solana_transaction_status_client_types

Core types for solana-transaction-status

Modules

option_serializer

Structs

ConfirmedTransactionStatusWithSignature

EncodedConfirmedBlock

EncodedConfirmedTransactionWithStatusMeta

EncodedTransactionWithStatusMeta

InnerInstruction

InnerInstructions

ParsedAccount

ParsedInstruction

Reward

TransactionStatus

TransactionStatusMeta

TransactionTokenBalance

UiAccountsList

UiAddressTableLookup

A duplicate representation of a MessageAddressTableLookup, in raw format, for pretty JSON serialization

UiCompiledInstruction

A duplicate representation of a CompiledInstruction for pretty JSON serialization

UiConfirmedBlock

UiInnerInstructions

UiLoadedAddresses

A duplicate representation of LoadedAddresses

UiParsedMessage

A duplicate representation of a Message, in parsed format, for pretty JSON serialization

UiPartiallyDecodedInstruction

A partially decoded CompiledInstruction that includes explicit account addresses

UiRawMessage

A duplicate representation of a Message, in raw format, for pretty JSON serialization

UiTransaction

A duplicate representation of a Transaction for pretty JSON serialization

UiTransactionError

UiTransactionReturnData

UiTransactionStatusMeta

A duplicate representation of TransactionStatusMeta with err field

UiTransactionTokenBalance

Enums

EncodeError

EncodedTransaction

ParsedAccountSource

TransactionBinaryEncoding

TransactionConfirmationStatus

TransactionDetails

UiInstruction

A duplicate representation of an Instruction for pretty JSON serialization

UiMessage

UiParsedInstruction

UiReturnDataEncoding

UiTransactionEncoding

Type Aliases

Rewards